// src/lib/tesserafin-sdk/generated/models/content-pack-browsing-preference.ts
/* tslint:disable */
/**
 * Tesserafin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * The way a user prefers the top level of the library to be arranged.
 * @export
 * @enum {string}
 */

export const ContentPackBrowsingPreference = {
    MediaFamilyFirst: 'MediaFamilyFirst',
    ContentPackFirst: 'ContentPackFirst'
} as const;

export type ContentPackBrowsingPreference = typeof ContentPackBrowsingPreference[keyof typeof ContentPackBrowsingPreference];